Rename export displacements to export vmf
It does brushes too now
