+++ addon/ops/__init__.py
@@ -6,7 +6,7 @@ from . compile_qc import SOURCEOPS_OT_CompileQC
 from . view_model import SOURCEOPS_OT_ViewModel
 from . open_log import SOURCEOPS_OT_OpenLog
 from . list_operator import SOURCEOPS_OT_ListOperator
-from . export_displacements import SOURCEOPS_OT_ExportDisplacements
+from . export_vmf import SOURCEOPS_OT_ExportVMF
 from . rig_simulation import SOURCEOPS_OT_RigSimulation
 
 
@@ -18,13 +18,13 @@ def register():
     bpy.utils.register_class(SOURCEOPS_OT_ViewModel)
     bpy.utils.register_class(SOURCEOPS_OT_OpenLog)
     bpy.utils.register_class(SOURCEOPS_OT_ListOperator)
-    bpy.utils.register_class(SOURCEOPS_OT_ExportDisplacements)
+    bpy.utils.register_class(SOURCEOPS_OT_ExportVMF)
     bpy.utils.register_class(SOURCEOPS_OT_RigSimulation)
 
 
 def unregister():
     bpy.utils.unregister_class(SOURCEOPS_OT_RigSimulation)
-    bpy.utils.unregister_class(SOURCEOPS_OT_ExportDisplacements)
+    bpy.utils.unregister_class(SOURCEOPS_OT_ExportVMF)
     bpy.utils.unregister_class(SOURCEOPS_OT_ListOperator)
     bpy.utils.unregister_class(SOURCEOPS_OT_OpenLog)
     bpy.utils.unregister_class(SOURCEOPS_OT_ViewModel)
